+++ CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+- [apm] fix: Add new option to `Tracing` `maxTransactionTimeout` determines the max length of a transaction
+- [hub] ref: Always also set transaction name on the top span in the scope
+
 ## 5.11.1
 
 - [apm] feat: Add build bundle including @sentry/browser + @sentry/apm